# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WildGuard proposed backend API",
    "version": "1.0.0",
    "description": "Original proposed full-system contract. Only milestone 1 is implemented; use implemented-openapi.json for its current envelope and APIs. Other modules remain future work."
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Planned full-system API; see implemented-openapi.json for the implemented subset."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/community-reports/{id}": {
      "put": {
        "tags": [
          "Community"
        ],
        "summary": "Submit community wildlife report",
        "operationId": "put_Community_29",
        "x-roles": [
          "COMMUNITY_MEMBER"
        ],
        "description": "Member has authorized park; server sets reporter and SUBMITTED. Require cropDetails for CROP_DAMAGE. Location coordinates may be omitted when only village/landmark is known. Any photo must be owned COMMUNITY_REPORT media in the same park. No SMS ingestion or automatic alert creation in this operation.",
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityReport"
                },
                "example": {
                  "id": "88888888-8888-4888-8888-888888888888",
                  "parkId": "park-yala",
                  "areaId": "area-north",
                  "type": "CROP_DAMAGE",
                  "species": "Elephant",
                  "village": "Village boundary Area B",
                  "occurredAt": "2026-10-08T13:00:00Z",
                  "description": "Elephants damaged the paddy field near the village.",
                  "cropDetails": "Paddy; damage near the northern edge of the field.",
                  "reportedBy": "usr-community-1",
                  "status": "SUBMITTED",
                  "createdAt": "2026-10-08T13:05:00Z"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "URI of the created resource.",
                "schema": {
                  "type": "string"
                },
                "example": "/api/v1/community-reports/88888888-8888-4888-8888-888888888888"
              }
            }
          },
          "200": {
            "description": "Identical replay returns the existing record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityReport"
                },
                "example": {
                  "id": "88888888-8888-4888-8888-888888888888",
                  "parkId": "park-yala",
                  "areaId": "area-north",
                  "type": "CROP_DAMAGE",
                  "species": "Elephant",
                  "village": "Village boundary Area B",
                  "occurredAt": "2026-10-08T13:00:00Z",
                  "description": "Elephants damaged the paddy field near the village.",
                  "cropDetails": "Paddy; damage near the northern edge of the field.",
                  "reportedBy": "usr-community-1",
                  "status": "SUBMITTED",
                  "createdAt": "2026-10-08T13:05:00Z"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "88888888-8888-4888-8888-888888888888"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommunityRequest"
              },
              "example": {
                "parkId": "park-yala",
                "areaId": "area-north",
                "type": "CROP_DAMAGE",
                "species": "Elephant",
                "village": "Village boundary Area B",
                "occurredAt": "2026-10-08T13:00:00Z",
                "description": "Elephants damaged the paddy field near the village.",
                "cropDetails": "Paddy; damage near the northern edge of the field."
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Community"
        ],
        "summary": "Get community report",
        "operationId": "get_Community_31",
        "x-roles": [
          "COMMUNITY_MEMBER",
          "PARK_MANAGER",
          "LIAISON_OFFICER"
        ],
        "description": "Owner member or authorized park manager/liaison only.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityReport"
                },
                "example": {
                  "id": "88888888-8888-4888-8888-888888888888",
                  "parkId": "park-yala",
                  "areaId": "area-north",
                  "type": "CROP_DAMAGE",
                  "species": "Elephant",
                  "village": "Village boundary Area B",
                  "occurredAt": "2026-10-08T13:00:00Z",
                  "description": "Elephants damaged the paddy field near the village.",
                  "cropDetails": "Paddy; damage near the northern edge of the field.",
                  "reportedBy": "usr-community-1",
                  "status": "SUBMITTED",
                  "createdAt": "2026-10-08T13:05:00Z"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "88888888-8888-4888-8888-888888888888"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": [
          "timestamp",
          "status",
          "code",
          "message",
          "path"
        ],
        "additionalProperties": false
      },
      "Location": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "minimum": -90,
            "maximum": 90
          },
          "longitude": {
            "type": "number",
            "minimum": -180,
            "maximum": 180
          },
          "source": {
            "type": "string",
            "enum": [
              "GPS",
              "MANUAL"
            ]
          },
          "accuracyMeters": {
            "type": "number",
            "minimum": 0,
            "maximum": 10000
          }
        },
        "required": [
          "latitude",
          "longitude",
          "source"
        ],
        "additionalProperties": false
      },
...
      "CommunityRequest": {
        "type": "object",
        "properties": {
          "parkId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": ".*\\S.*"
          },
          "areaId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": ".*\\S.*"
          },
          "type": {
            "type": "string",
            "enum": [
              "WILDLIFE_SIGHTING",
              "CROP_DAMAGE"
            ]
          },
          "species": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "pattern": ".*\\S.*"
          },
          "village": {
            "type": "string",
            "minLength": 1,
            "maxLength": 150,
            "pattern": ".*\\S.*"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000,
            "pattern": ".*\\S.*"
          },
          "cropDetails": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500,
            "pattern": ".*\\S.*"
          },
          "photoId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "parkId",
          "areaId",
          "type",
          "village",
          "occurredAt",
          "description"
        ],
        "additionalProperties": false
      },
      "CommunityReport": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "parkId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": ".*\\S.*"
          },
          "areaId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": ".*\\S.*"
          },
          "type": {
            "type": "string",
            "enum": [
              "WILDLIFE_SIGHTING",
              "CROP_DAMAGE"
            ]
          },
          "species": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "pattern": ".*\\S.*"
          },
          "village": {
            "type": "string",
            "minLength": 1,
            "maxLength": 150,
            "pattern": ".*\\S.*"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000,
            "pattern": ".*\\S.*"
          },
          "cropDetails": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500,
            "pattern": ".*\\S.*"
          },
          "photoId": {
            "type": "string",
            "format": "uuid"
          },
          "reportedBy": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": ".*\\S.*"
          },
          "status": {
            "type": "string",
            "enum": [
              "SUBMITTED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "parkId",
          "areaId",
          "type",
          "village",
          "occurredAt",
          "description",
          "reportedBy",
          "status",
          "createdAt"
        ],
        "additionalProperties": false
      },
...
    },
    "responses": {
      "Error400": {
        "description": "Invalid request or validation failed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 400,
              "code": "VALIDATION_FAILED",
              "message": "Invalid request or validation failed.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error401": {
        "description": "Missing/invalid token, inactive account or invalid login.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 401,
              "code": "UNAUTHORIZED",
              "message": "Missing/invalid token, inactive account or invalid login.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error403": {
        "description": "Role or requested park is forbidden.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 403,
              "code": "FORBIDDEN",
              "message": "Role or requested park is forbidden.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error404": {
        "description": "Resource not found or not accessible.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 404,
              "code": "NOT_FOUND",
              "message": "Resource not found or not accessible.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error409": {
        "description": "Conflicting state or resource ID reused with different data.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 409,
              "code": "STATE_CONFLICT",
              "message": "Conflicting state or resource ID reused with different data.",
              "path": "/api/v1/example"
            }
          }
        }
      },
...
      "Error500": {
        "description": "Unexpected server error; no false success acknowledgement.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 500,
              "code": "INTERNAL_ERROR",
              "message": "Unexpected server error; no false success acknowledgement.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error503": {
        "description": "Database/storage temporarily unavailable; retry later.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 503,
              "code": "SERVICE_UNAVAILABLE",
              "message": "Database/storage temporarily unavailable; retry later.",
              "path": "/api/v1/example"
            }
          }
        }
      }
    }
  }
}